consolidate to generic.py
