+++ pandas/core/groupby/__init__.py
@@ -1,5 +1,4 @@
-# flake8: noqa
-from pandas.core.groupby.groupby import GroupBy
-from pandas.core.groupby.series import SeriesGroupBy
-from pandas.core.groupby.frame import DataFrameGroupBy
-from pandas.core.groupby.ops import Grouper
+from pandas.core.groupby.groupby import GroupBy  # flake8: noqa
+from pandas.core.groupby.generic import (
+    SeriesGroupBy, DataFrameGroupBy, PanelGroupBy)  # flake8: noqa
+from pandas.core.groupby.ops import Grouper  # flake8: noqa
