+++ pandas/core/groupby/base.py
@@ -3,8 +3,8 @@ base components for groupby
 """
 
 import types
-from pandas.core.dtypes.common import is_scalar, is_list_like
 from pandas.util._decorators import make_signature
+from pandas.core.dtypes.common import is_scalar, is_list_like
 
 
 class GroupByMixin(object):
